% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{changes_over_expac}
\alias{changes_over_expac}
\title{Outputs an interactive animated plot that displays the count of card for each expansion in H}
\usage{
changes_over_expac(df_cards = get_all_cards(), classifier = "MinionType")
}
\arguments{
\item{df_cards}{optional argument. if you already have a data frame generated from any of the get_* functions you may pass it so the function does not have to load a new one. Default set to automatically generate data from get_all_cards function}

\item{classifier}{a string object that determines what the plot will output, must be "MinionType", "Rarity, or "CardType"}
}
\value{
An interactive plotly object that allows the user to view the distribution of the cards released for the selected expansion
}
\description{
Outputs an interactive animated plot that displays the count of card for each expansion in H
}
