% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_all_cards}
\alias{get_all_cards}
\title{Returns a neat data frame of information about all hearthstone cards of a desired filter}
\usage{
get_all_cards(class = NULL, mana = NULL)
}
\arguments{
\item{class}{a string filters the cards that return to only be from a specific class in Hearthstone, default is NULL to return all Classes}

\item{mana}{a string argument, filters the data frame to only include cards of a certain mana cost. Default is NULL which returns cards of all mana costs.}
}
\value{
returns a data frame with Name of card, Class, Type, Rarity, Attack, Health, what set it was apart of, durability, armor, and length of card text
}
\description{
Returns a neat data frame of information about all hearthstone cards of a desired filter
}
