% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_decks}
\alias{get_decks}
\title{Returns data frame of cards in a deck using given in game generated deck code}
\usage{
get_decks(deckcode)
}
\arguments{
\item{deckcode}{a string of characters that identifies a deck in the game of hearthstone}
}
\value{
returns a data frame with with all the cards for a given
}
\description{
Returns data frame of cards in a deck using given in game generated deck code
}
