% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_minions}
\alias{get_minions}
\title{main difference between get_minions and get_all_cards is that it returns a data frame with variables only containing values pertaining to minions}
\usage{
get_minions(class = NULL, mana = NULL, attack = NULL, health = NULL)
}
\arguments{
\item{class}{a string that filters results to only show specific class in Hearthstone, default is NULL to return all Classes. MUST BE PASSED AS FOLLOWS: "demonhunter" "druid","hunter","mage","paladin","priest","rogue","shaman","warlock", or "warrior"}

\item{mana}{a string argument, filters the data frame to only include cards of a certain mana cost. Default is NULL which returns cards of all mana costs.}

\item{attack}{a string argument, filters the data frame to only include cards with a specific amount of attack. Default is NULL which returns cards of all attack strengths.}

\item{health}{a string argument, filters the data frame to only include cards with a specific amount of health. Default is NULL which returns cards of all health totals}
}
\value{
returns a data frame with Name of card, Class, Type, Rarity, what set it was apart of, and length of card text
}
\description{
main difference between get_minions and get_all_cards is that it returns a data frame with variables only containing values pertaining to minions
}
