% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_class_stats}
\alias{plot_class_stats}
\title{Compare class card statistics between your favorite heroes and villains of Azeroth.}
\usage{
plot_class_stats(
  df_cards = NULL,
  comparison_var,
  cardType = "minion",
  animate = FALSE
)
}
\arguments{
\item{df_cards}{optional argument. if you already have a data frame generated from any of the \verb{get_*} functions you may pass it so the function does not have to load a new one. default is null and will select best \verb{get_*} function for your query.}

\item{comparison_var}{string variable argument. Selects a variable to compare the classes by using the get_* functions of the hearrthstone package, get_all_cards for details}

\item{cardType}{string argument of what data you would like to compare the classes by, minions, spells, weapons, or all types. To compare all types pass NULL. Default set to minions.}

\item{animate}{animate will allow return an animated plot changing the comparison by the various card sets in the game.}
}
\value{
Default outputs an interactive plotly object or a gif if the animate argument is passed as TRUE
}
\description{
Compare class card statistics between your favorite heroes and villains of Azeroth.
}
