% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_rarity_stats}
\alias{plot_rarity_stats}
\title{Compare different card rarities to see if legendary and epic cards are really worth your hard earned dust}
\usage{
plot_rarity_stats(
  df_cards = NULL,
  comparison_var,
  cardType = "Minion",
  class = NULL,
  animate = FALSE
)
}
\arguments{
\item{df_cards}{optional argument. if you already have a data frame generated from any of the \verb{get_*} functions you may pass it so the function does not have to load a new one. default is null and will select best \verb{get_*} function for your query.}

\item{comparison_var}{string variable argument. Selects a variable to compare the classes by using the get_* functions of the hearrthstone package, get_all_cards for details}

\item{cardType}{string argument of what data you would like to compare the classes by, minions, spells, weapons, or all types. To compare all types pass NULL. Default set to minions.}

\item{class}{a string that filters results to only show a specific class in Hearthstone, default is NULL to return all Classes. MUST BE PASSED AS FOLLOWS: "demonhunter" "druid","hunter","mage","paladin","priest","rogue","shaman","warlock", "warrior", "neutral"}

\item{animate}{animate will allow return an animated plot changing the comparison by the various card sets in the game.}
}
\value{
Default outputs an interactive plotly object or a gif if the animate argument is passed as TRUE
}
\description{
note: there are no neutral spells OR weapons
}
