% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{token_auth}
\alias{token_auth}
\title{Regenerate Token}
\usage{
token_auth()
}
\value{
an access token to be used for all other functions
}
\description{
The blizzard API is a little funky and requires you to regenerate an access token every 24 hours.
This function allows us to regenerate the keys within our functions so we don't have to worry about that again
For the purposes of this package we will be using my developer key and client secret.
}
